requirements.txt
asgiref
astunparse
blinker
certifi
charset-normalizer
click
django
flask
flatbuffers
gast
google-pasta
grpcio
gunicorn
h5py
idna
itsdangerous
jinja2
keras
keras-preprocessing
markdown
markdown-it-py
markupsafe
mdurl
ml-dtypes
namex
numpy
opt-einsum
optree
packaging
pillow
protobuf
psycopg2-binary
pygments
requests
rich
six
sqlparse
tensorboard
tensorboard-data-server
tensorflow
tensorflow-cpu
termcolor
typing-extensions
urllib3
werkzeug
wrapt
flask-cors
